Fix convetional tts structure and add licences.
