+++ nemo/collections/tts/__init__.py
@@ -14,8 +14,8 @@
 # =============================================================================
 
 from nemo.collections.tts.data_layers import AudioDataLayer
-from nemo.collections.tts.fastspeech import *
-from nemo.collections.tts.fastspeech import __all__ as fastspeech__all__
+from nemo.collections.tts.fastspeech_modules import *
+from nemo.collections.tts.fastspeech_modules import __all__ as fastspeech__all__
 from nemo.collections.tts.parts.helpers import *
 from nemo.collections.tts.parts.helpers import __all__ as helpers__all__
 from nemo.collections.tts.tacotron2_modules import *
