--- nemo/collections/tts/fastspeech/__init__.py
@@ -1,9 +1,0 @@
-from nemo.collections.tts.fastspeech.data import FastSpeechDataLayer
-from nemo.collections.tts.fastspeech.loss import FastSpeechLoss
-from nemo.collections.tts.fastspeech.model import FastSpeech
-
-__all__ = [
-    'FastSpeechDataLayer',
-    'FastSpeechLoss',
-    'FastSpeech',
-]
