--- nemo/collections/tts/fastspeech/transformer/__init__.py
@@ -1,2 +1,0 @@
-from nemo.collections.tts.fastspeech.transformer.conv_transpose import Conv
-from nemo.collections.tts.fastspeech.transformer.models import Decoder, Encoder
