+++ nemo/collections/tts/parts/transformer/layers.py
@@ -1,10 +1,24 @@
+# Copyright 2020 NVIDIA. All Rights Reserved.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
 from collections import OrderedDict
 
 import torch
 from torch import nn
 from torch.nn import functional as F
 
-from nemo.collections.tts.fastspeech.transformer import sublayers
+from nemo.collections.tts.parts.transformer import sublayers
 
 
 class Linear(nn.Module):
