+++ nemo/collections/tts/parts/transformer/models.py
@@ -1,8 +1,22 @@
+# Copyright 2020 NVIDIA. All Rights Reserved.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
 import numpy as np
 import torch
 from torch import nn
 
-from nemo.collections.tts.fastspeech.transformer import layers
+from nemo.collections.tts.parts.transformer import layers
 
 
 def get_non_pad_mask(seq, pad_id):
@@ -11,7 +25,7 @@ def get_non_pad_mask(seq, pad_id):
 
 
 def get_sinusoid_encoding_table(n_position, d_hid, padding_idx=None):
-    ''' Sinusoid position encoding table '''
+    """Sinusoid position encoding table."""
 
     def cal_angle(position, hid_idx):
         return position / np.power(10000, 2 * (hid_idx // 2) / d_hid)
@@ -32,7 +46,7 @@ def get_sinusoid_encoding_table(n_position, d_hid, padding_idx=None):
 
 
 def get_attn_key_pad_mask(seq_k, seq_q, pad_id):
-    ''' For masking out the padding part of key sequence. '''
+    """For masking out the padding part of key sequence."""
 
     # Expand to fit the shape of key query attention matrix.
     len_q = seq_q.size(1)
