+++ setup.py
@@ -28,6 +28,19 @@ from itertools import chain
 
 import setuptools
 
+from nemo.package_info import (
+    __contact_emails__,
+    __contact_names__,
+    __description__,
+    __download_url__,
+    __homepage__,
+    __keywords__,
+    __license__,
+    __package_name__,
+    __repository_url__,
+    __version__,
+)
+
 
 def is_build_action():
     if len(sys.argv) <= 1:
@@ -43,19 +56,6 @@ def is_build_action():
 
 if is_build_action():
     os.environ['NEMO_PACKAGE_BUILDING'] = 'True'
-
-from nemo.package_info import (
-    __contact_emails__,
-    __contact_names__,
-    __description__,
-    __download_url__,
-    __homepage__,
-    __keywords__,
-    __license__,
-    __package_name__,
-    __repository_url__,
-    __version__,
-)
 
 
 if os.path.exists('nemo/README.md'):
